Rename package. Move files
