+++ generator/src/main/java/com/scottlogic/deg/generator/generation/string/generators/ChecksumStringGeneratorFactory.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.scottlogic.deg.generator.generation.string.streamy;
+package com.scottlogic.deg.generator.generation.string.generators;
 
-import com.scottlogic.deg.generator.generation.string.RegexStringGenerator;
 import com.scottlogic.deg.generator.utils.FinancialCodeUtils;
 
 public class ChecksumStringGeneratorFactory {
